Extracts helper methods to helpers.rb
