+++ app/app.rb
@@ -7,6 +7,7 @@ require 'sinatra/partial'
 require_relative './data_mapper_setup'
 
 require_relative 'server'
+require_relative 'helpers'
 require_relative 'controllers/navigation'
 require_relative 'controllers/peeping'
 require_relative 'controllers/users'
